simplint1.0

+++ simplint.c
@@ -2,18 +2,18 @@
 #include<stdio.h>
 int main()
 {
-    int principle;
+    float principle;
     float rate;
-    int time;
+    float time;
     float simpleint;
     simpleint=principle*rate*time;
     printf("Enter Principle amount:");
-    scanf("%d",&principle);
+    scanf("%f",&principle);
     printf("Enter Rate:");
     scanf("%f",&rate);
     printf("Enter Time:");
-    scanf("%d",&time);
+    scanf("%f",&time);
     simpleint=(principle*rate*time)/100;
-    printf("Simpele Interest is %0.0f",simpleint);
+    printf("Simpele Interest is %0.2f",simpleint);
     return 0;
 }
